refactor: fetch 로직 apis로 옮김
