+++ frontend/src/apis/auth/types.ts
@@ -1,0 +1,8 @@
+interface TokenValidResponse {
+  id: number;
+  email: string;
+  nickname: string;
+  createdAt: string;
+  updatedAt: string;
+}
+export { type TokenValidResponse };
